[ENG-588] Use insert on duplicate key update.
This is necessary because the $changes array is not always trustable.
Enhancements validations, normalizations may occur causing the changes
array to not be indicative of the difference between what is in the
database and what is in memory. As a result, we are switching to an
insert on duplicate key update pattern, which will fail far less often
and ensure we are always persisting the most recent value.